projection: add video output that use Gstreamer and QtQuick
This is due to QtMultimedia on Android not supporting playing from
QIODevice and QtMultimediaWidget not available on Android.

So Gstreamer is choosen as a video player, as it has support for Android
video decoding API and can render to QtQuick. Unfortunately this means
the app will now pull in QtQuick and QtQml in addition to QtWidget, but
this will allow us to do more things in the future e.g. automatically
add video margin based on actual video output area to not distort video
on non 16:9 display.